✨ 카카오 로그인 기능 구현

+++ data/src/main/java/com/wonder/data/repository/AuthRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.wonder.data.repository
 
 import com.wonder.data.remote.auth.AuthRemoteDataSource
-import com.wonder.data.service.AuthService
 import com.wonder.domain.model.auth.UserAuthInfo
 import com.wonder.domain.repository.AuthRepository
 import javax.inject.Inject
@@ -10,5 +9,6 @@ internal class AuthRepositoryImpl @Inject constructor(
     private val authRemoteDataSource: AuthRemoteDataSource
 ) : AuthRepository {
 
-    override suspend fun signInKakao(uid: Int): UserAuthInfo = authRemoteDataSource.signInKakao(uid)
+    override suspend fun signInKakao(uid: Long): UserAuthInfo =
+        authRemoteDataSource.signInKakao(uid)
 }
